BottomNavigationBar icon enabled
